+++ apps/docs/src/pages/_app.tsx
@@ -1,0 +1,11 @@
+import { ReCompUIProvider } from '@recompui/core';
+
+function MyApp({ Component, pageProps }) {
+  return (
+    <ReCompUIProvider>
+      <Component {...pageProps} />
+    </ReCompUIProvider>
+  );
+}
+
+export default MyApp;