+++ packages/recompui-core/src/theme/index.ts
@@ -1,0 +1,4 @@
+export * from './colors';
+export * from './stitches.config';
+export { default as ReCompUIProvider } from './theme-provider';
+export * from './types';